==> R/00_doit.R <==
library(tidyverse)
library(magrittr)
library(ComplexHeatmap)
library(janitor)
library(ggfortify)
library(plyr)
library(ggrepel)
library(readxl)
library(patchwork)

source("R/99_proj_func.R")
source("R/01_load.R")
source("R/02_clean.R")
source("R/03_augment.R")
source("R/04_metadata_overview.R")
source("R/05_pca.R")
source("R/06_volcanoplot.R")
source("R/07_heatmap_immuno.R")
source("R/additional_plots.R")
source("R/neutralization_analysis.R")
